requirements.txt
discord.py
google-generativeai
python-dotenv
google-api-python-client
google-auth
